renamed debug to print and added types.cpp to git
